url-shortener/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
passlib[argon2]
argon2-cffi
python-jose[cryptography]
python-multipart
pytest
pytest-asyncio
httpx
